requirements.txt
streamlit
scikit-learn
numpy
pandas
matplotlib
seaborn
joblib
pillow
